prefcenter package ID updates
