+++ ketch-android/Android Ketch Pref Center using Kotlin/app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdk 33
 
     defaultConfig {
-        applicationId "com.ketch.sample.prefcenter"
+        applicationId "com.ketch.sample.pref"
         minSdk 21
         targetSdk 33
         versionCode 1
