+++ ketch-android/Android Ketch Pref Center using Kotlin/app/src/main/java/com/ketch/sample/pref/Identity.kt
@@ -1,0 +1,7 @@
+package com.ketch.sample.pref
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Identity(val code: String, val value: String) : Parcelable
