--- ketch-android/Android Ketch Pref Center using Kotlin/app/src/main/java/com/ketch/sample/prefcenter/Identity.kt
@@ -1,7 +1,0 @@
-package com.ketch.sample.prefcenter
-
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-
-@Parcelize
-data class Identity(val code: String, val value: String) : Parcelable
