+++ ketch-android/Android Ketch Pref Center using Kotlin/app/src/main/res/layout/activity_main.xml
@@ -41,7 +41,7 @@
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 
-    <com.ketch.sample.prefcenter.KetchWebView
+    <com.ketch.sample.pref.KetchWebView
         android:id="@+id/ketchWebView"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
